% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{add_interaction_type}
\alias{add_interaction_type}
\title{Add interaction type to static or dynamic GRN}
\usage{
add_interaction_type(grn)
}
\arguments{
\item{grn}{a static or dynamic grn}
}
\value{
an updated static or dynamic grn with added interaction type
}
\description{
Add interaction type to static or dymamic GRN
}
