% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkComparison.R
\name{add_type}
\alias{add_type}
\title{Adds interaction type to dynamic differential network}
\usage{
add_type(diffnet, type, grnDF_on, grnDF_offlist)
}
\arguments{
\item{type}{"on" or "off" depending on the type of differential network. If "on" will assign type based on grnDF_on. Otherwise interaction assigned from grnDF_offlist.}

\item{grnDF_on}{the static network in which diffnet edges are active}

\item{grnDF_offlist}{list of static networks in which diffnet edges are inactive}
}
\value{
community assignments of nodes in the dynamic network
}
\description{
Adds interaction type to dynamic differential network
}
