% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{assign_epochs}
\alias{assign_epochs}
\title{Assigns genes to epochs}
\usage{
assign_epochs(
  expDat,
  dynRes,
  method = "active_expression",
  pThresh_dyn = 0.05,
  pThresh_DE = 0.05,
  active_thresh = 0.33,
  toScale = FALSE,
  forceGenes = TRUE
)
}
\arguments{
\item{expDat}{genes-by-cells expression matrix}

\item{dynRes}{individual path result of running define_epochs}

\item{method}{method of assigning epoch genes, either "active_expression" (looks for active expression in epoch) or "DE" (looks for differentially expressed genes per epoch)}

\item{pThresh_dyn}{pval threshold if gene is dynamically expressed}

\item{pThresh_DE}{pval if gene is differentially expressed. Ignored if method is active_expression.}

\item{active_thresh}{value between 0 and 1. Percent threshold to define activity}

\item{toScale}{whether or not to scale the data}

\item{forceGenes}{whether or not to rescue orphan dyanmic genes, forcing assignment into epoch with max expression.}
}
\value{
epochs a list detailing genes active in each epoch
}
\description{
Assigns genes to epochs
}
