% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{assign_epochs_simple}
\alias{assign_epochs_simple}
\title{Assigns genes to epochs just based on which mean is maximal}
\usage{
assign_epochs_simple(
  expDat,
  dynRes,
  num_epochs = 3,
  pThresh = 0.01,
  toScale = FALSE
)
}
\description{
Assigns genes to epochs just based on which mean is maximal
}
