% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkComparison.R
\name{biglist_compute_betweenness_degree}
\alias{biglist_compute_betweenness_degree}
\title{Computes betweenness and degree of each TF for each network in a list of networks}
\usage{
biglist_compute_betweenness_degree(
  netlist,
  weight_column = "zscore",
  normalized = TRUE
)
}
\arguments{
\item{weight_column}{column name in grnDFs containing edge weights}

\item{normalized}{whether or not to normalize degree and betweenness}
}
\value{
dataframe listing network, betweenness of each TF, degree of each TF
}
\description{
Computes betweenness and degree of each TF for each network in a list of networks
}
