% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ButterworthFilter.R
\name{butterworthFilter}
\alias{butterworthFilter}
\title{Apply Fourth-Order Butterworth Filter to Epoch Data}
\usage{
butterworthFilter(epoch, lowpass = NULL, highpass = NULL, order = 4)
}
\arguments{
\item{epoch}{An Epoch object containing iEEG data}

\item{lowpass}{Numeric. Lowpass cutoff frequency in Hz. The default is 0.5 Hz.}

\item{highpass}{Numeric. Highpass cutoff frequency in Hz. The default is 99\% of the Nyquist frequency.}

\item{order}{Integer. The order of the Butterworth filter. Default is 4.}
}
\value{
An Epoch object with filtered data
}
\description{
Apply a fourth-order Butterworth filter to iEEG epoch data.
This function can apply lowpass, highpass, or bandpass filtering depending
on the parameters provided.
}
