% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{caoGenes}
\alias{caoGenes}
\title{cluster and order genes}
\usage{
caoGenes(expSmooth, expDat, dynRes, k = 3, pThresh = 0.01, method = "kmeans")
}
\arguments{
\item{expSmooth}{expression matrix}

\item{dynRes}{result of running findDynGenes}

\item{k}{number of clusters}

\item{pThresh}{pval threshold}
}
\value{
data.frame of dynamically expressed genes, cluster, peakTime, ordered by peaktime
}
\description{
cluster and order genes
}
