% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_GRN.R
\name{cn_extractRegsDF}
\alias{cn_extractRegsDF}
\title{Create GRN table}
\usage{
cn_extractRegsDF(zscores, corrMatrix, genes, threshold, removeAuto = TRUE)
}
\value{
data frame of GRN
}
\description{
Create GRN table
}
