% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{compile_epochs}
\alias{compile_epochs}
\title{compile epochs on branched trajectories}
\usage{
compile_epochs(epochlist)
}
\arguments{
\item{epochlist}{list of results of running assign_epoch on each path}
}
\value{
a compiled list of epoch genes
}
\description{
compile epochs based on matching epoch label in result of running assign_epoch
}
