% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{compileDynGenes}
\alias{compileDynGenes}
\title{compiles results from findDynGenes run on separate paths}
\usage{
compileDynGenes(dynReslist)
}
\arguments{
\item{dynReslist}{list of results from findDynGenes}
}
\value{
a new dynRes a.k.a. the compiled result of findDynGenes
}
\description{
compiles results from findDynGenes run on separate paths into a single result
}
