% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{compute_betweenness_degree}
\alias{compute_betweenness_degree}
\title{Function to compute betweenness and degree}
\usage{
compute_betweenness_degree(
  dynnet,
  weight_column = "zscore",
  directed_graph = FALSE
)
}
\arguments{
\item{dynnet}{result of dynamic GRN reconstruction}

\item{weight_column}{name of column in dynnet to weight edges}

\item{directed_graph}{if GRN is directed or not}
}
\value{
list of dataframes of active TFs in each epoch and transition, ranked by betweenness*degree
}
\description{
Function to compute betweenness and degree
}
