% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkComparison.R
\name{compute_frobenius_distance}
\alias{compute_frobenius_distance}
\title{Computes frobenius distance in a pairwise manner between two sets of networks}
\usage{
compute_frobenius_distance(
  netlist1,
  netlist2,
  weight_column = "zscore",
  compare_within_netlist1 = TRUE,
  compare_within_netlist2 = TRUE
)
}
\arguments{
\item{netlist1}{list of grnDFs}

\item{netlist2}{list of grnDFs}

\item{weight_column}{column name containing edge weights}

\item{compare_within_netlist1}{whether or not to do pairwise comparisons between networks in netlist1}

\item{compare_within_netlist2}{whether or not to do pairwise comparisons between networks in netlist2}
}
\value{
dataframe of frobenius distances
}
\description{
Computes frobenius distance in a pairwise manner between two sets of networks
}
