% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{compute_pagerank}
\alias{compute_pagerank}
\title{Function to compute page rank of TF+target networks}
\usage{
compute_pagerank(dynnet, weight_column = "zscore", directed_graph = FALSE)
}
\arguments{
\item{dynnet}{result of dynamic GRN reconstruction}

\item{weight_column}{name of column in dynnet to weight edges}

\item{directed_graph}{if GRN is directed or not}
}
\value{
list of dataframes of active genes in each epoch and transition, ranked by page rank
}
\description{
Function to compute page rank of TF+target networks
}
