% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_crossweighting.R
\name{crossweight}
\alias{crossweight}
\title{Perform crossweighting}
\usage{
crossweight(
  grnTab,
  expDat,
  xdyn,
  lag = floor(ncol(expDat)/5),
  min = ceiling(ncol(expDat)/50),
  max = floor(ncol(expDat)/12),
  symmetric_filter = FALSE,
  filter_thresh = 0
)
}
\arguments{
\item{grnTab}{GRN dataframe, the result of running reconstructGRN or reconstructGRN_GENIE3}

\item{expDat}{genes-by-cells expression matrix}

\item{xdyn}{result of running findDynGenes}

\item{lag}{lag window on which to run cross-correlation. Cross-correlaiton computed from -lag to +lag.}

\item{min}{minimum of weighting window. Edges with offsets (or absolute offsets if symmetric_filter=TRUE) less than min will not be negatively weighted.}

\item{max}{maximum of weighting window. Edges with offsets (or absolute offsets if symmetric_filter=TRUE) greater than max will have weights set to 0.}

\item{symmetric_filter}{whether or not to employ a symmetric weight scheme. If true, absolute offset is used in place of offset.}

\item{filter_thresh}{after crossweighting, edges with weights less than filter_thresh will be set to 0.}
}
\value{
grnTab with offset and weighted_score added
}
\description{
Perform crossweighting
}
