% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{define_epochs}
\alias{define_epochs}
\title{Define epochs}
\usage{
define_epochs(
  dynRes,
  expDat,
  method = "pseudotime",
  num_epochs = 2,
  pseudotime_cuts = NULL,
  group_assignments = NULL,
  pThresh_dyn = 0.05,
  winSize = 2
)
}
\arguments{
\item{dynRes}{results of running findDynGenes, or a list of results of running findDynGenes per path. If list, names should match names of expDat.}

\item{expDat}{genes-by-cells expression matrix, or a list of expression matrices per path. If list, names should match names of dynRes.}

\item{method}{method to define epochs. Either "pseudotime", "cell_order", "group", "con_similarity", "kmeans", "hierarchical"}

\item{num_epochs}{number of epochs to define. Ignored if epoch_transitions, pseudotime_cuts, or group_assignments are provided.}

\item{pseudotime_cuts}{vector of pseudotime cutoffs. If NULL, cuts are set to max(pseudotime)/num_epochs.}

\item{group_assignments}{a list of vectors where names(assignment) are epoch names, and vectors contain groups belonging to corresponding epoch}
}
\value{
updated list of dynRes with epoch column included in dynRes$cells
}
\description{
Define epochs
}
