% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkComparison.R
\name{diffnet_community_detection}
\alias{diffnet_community_detection}
\title{Perform community detection on a dynamic network}
\usage{
diffnet_community_detection(
  diffnet,
  method = "louvain",
  use_weights = FALSE,
  weight_column = NULL
)
}
\arguments{
\item{method}{community detection method, currently only "louvain"}

\item{use_weights}{whether or not to use edge weights (for weighted graphs)}

\item{weight_column}{if using weights, name of the column containing edge weights}
}
\value{
community assignments of nodes in the dynamic network
}
\description{
Perform community detection on a dynamic network
}
