% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkComparison.R
\name{dynamic_difference_network}
\alias{dynamic_difference_network}
\title{Compute a dynamic difference network}
\usage{
dynamic_difference_network(
  edgeDF,
  epochs,
  condition,
  type,
  diff_thresh = 3,
  condition_thresh = 6
)
}
\arguments{
\item{edgeDF}{the result of running edge_uniqueness}

\item{epochs}{list of epoch gene assignments}

\item{condition}{condition of interest, should be one of the treatment (aka network name) or column names in edgeDF}

\item{type}{"on" or "off" specifies either finding edges that are active in the condition network but off others or inactive in condition network but active in others}

\item{diff_thresh}{edge difference threshold to determine if edge is uniquely on or off}

\item{condition_thresh}{edge weight theshold in condition network to keep or filter out in difference network}
}
\value{
list of dataframes representing the dynamic difference network
}
\description{
Compute a dynamic difference network
}
