% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_networkAnalysis.R
\name{dynamic_shortest_path}
\alias{dynamic_shortest_path}
\title{Function to return shortest path from 1 TF to 1 TG in a dynamic network}
\usage{
dynamic_shortest_path(
  grn,
  from,
  to,
  weight_column = "weighted_score",
  compare_to_average = FALSE
)
}
\arguments{
\item{from}{the starting TF}

\item{to}{the end TF}

\item{weight_column}{column name in grnDF with edge weights that will be converted to distances}

\item{compare_to_average}{if TRUE will compute normalized against average path length}

\item{grnDF}{a dyanmic network}
}
\value{
shortest path, distance, normalized distance, and action
}
\description{
Function to return shortest path from 1 TF to 1 TG in a dynamic network
}
