% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_networkAnalysis.R
\name{dynamic_shortest_path_multiple}
\alias{dynamic_shortest_path_multiple}
\title{Function to return shortest path from multiple TFs to multiple targets in a dynamic network}
\usage{
dynamic_shortest_path_multiple(grn, from, to, weight_column = "weighted_score")
}
\arguments{
\item{from}{the starting TFs}

\item{to}{the end TFs}

\item{weight_column}{column name in grnDF with edge weights that will be converted to distances}

\item{grnDF}{a dyanmic network}
}
\value{
dataframe with shortest path, distance, normalized distance, and action
}
\description{
Function to return shortest path from multiple TFs to multiple targets in a dynamic network
}
