% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkComparison.R
\name{edge_rank}
\alias{edge_rank}
\title{Function to compute edge differences}
\usage{
edge_rank(grnMats)
}
\arguments{
\item{grnMats}{list of adjacency matrices}
}
\value{
dataframe of edges with difference scores between the GRNs
}
\description{
Function to compute edge differences
}
