% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkComparison.R
\name{edge_uniqueness}
\alias{edge_uniqueness}
\title{Function to compute edge differences between networks}
\usage{
edge_uniqueness(grnDFs, tfs, weight_column)
}
\arguments{
\item{grnDFs}{list of static grnDFs}

\item{tfs}{TFs}

\item{weight_column}{column name in grnDFs containing edge weight}
}
\value{
dataframe of edges with difference scores between the GRNs
}
\description{
Function to compute edge differences between networks
}
