% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Epoch.r, R/method-print.r,
%   R/method-wrappers.r
\name{crop}
\alias{crop}
\alias{crop,Epoch-method}
\alias{coltimes}
\alias{coltimes,Epoch-method}
\alias{show,Epoch-method}
\alias{dim,Epoch-method}
\alias{dimnames,Epoch-method}
\alias{[,Epoch-method}
\title{Methods for Epoch class}
\usage{
crop(x, start, end)

\S4method{crop}{Epoch}(x, start, end)

coltimes(x)

\S4method{coltimes}{Epoch}(x)

\S4method{show}{Epoch}(object)

\S4method{dim}{Epoch}(x)

\S4method{dimnames}{Epoch}(x)

\S4method{[}{Epoch}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An Epoch object}

\item{start}{Numeric value specifying start of new time range}

\item{end}{Numeric value specifying end of new time range}

\item{object}{Epoch object}

\item{i}{Row indices for subsetting. If only \code{i} is provided, it will return the entire row(s).}

\item{j}{Column indices for subsetting.}

\item{...}{Additional arguments.}

\item{drop}{Not used.}
}
\value{
clip: clip the time range of the Epoch object

\code{coltimes}: A numeric vector of time points, or column indices if time points are not defined

\code{[}: A new Epoch object with the selected data.
}
\description{
\code{clip}: Truncating time range
}
