% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.r
\docType{class}
\name{EpochDownloader-class}
\alias{EpochDownloader-class}
\alias{.EpochDownloader}
\title{EpochDownloader}
\description{
\code{EpochDownloader} is a class that allows
downloading and accessing files from a OSF project.
}
\section{Slots}{

\describe{
\item{\code{id}}{The ID of the OSF project.}

\item{\code{files}}{The files in the OSF project.}

\item{\code{dataNames}}{The names of the files in the OSF project.}

\item{\code{tmp_folder}}{The temporary folder where the files are downloaded.}

\item{\code{progress}}{Logical indicating whether to show progress during download.}
}}

