% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.r
\name{EpochDownloader}
\alias{EpochDownloader}
\title{EpochDownloader constructor}
\usage{
EpochDownloader(id = "7rnft", progress = TRUE, path = file.path(tempdir(), id))
}
\arguments{
\item{id}{The ID of the OSF project (the default points to the BHI project).}

\item{progress}{Logical indicating whether to show progress during download.}

\item{path}{The path to the temporary folder where the files will be downloaded.}
}
\value{
An \code{EpochDownloader} object.
}
\description{
EpochDownloader constructor
}
