% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{epochGRN}
\alias{epochGRN}
\title{Divides grnDF into epochs, filters interactions between genes not in same or consecutive epochs}
\usage{
epochGRN(grnDF, epochs, epoch_network = NULL)
}
\arguments{
\item{grnDF}{result of GRN reconstruction}

\item{epochs}{result of running assign_epochs}

\item{epoch_network}{dataframe outlining higher level epoch connectivity (i.e. epoch transition network). If NULL, will assume epochs is ordered linear trajectory}
}
\value{
list of GRNs across epochs and transitions
}
\description{
Divides grnDF into epochs, filters interactions between genes not in same or consecutive epochs
}
