% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpochRepos.r
\name{EpochRepos}
\alias{EpochRepos}
\title{Get the list of available projects}
\usage{
EpochRepos(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical indicating whether to show messages}
}
\value{
A list of project names and their corresponding OSF project IDs.
}
\description{
This function returns the list of available projects.
If the configuration has not been fetched yet, it will
automatically update the project list from the remote repository.
}
\examples{
EpochRepos()
}
