% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{find_cuts_by_clustering}
\alias{find_cuts_by_clustering}
\title{Returns cuts to define epochs}
\usage{
find_cuts_by_clustering(
  expDat,
  dynRes,
  num_epochs,
  limit_to = NULL,
  method = "kmeans",
  pThresh_dyn = 0.05
)
}
\arguments{
\item{expDat}{genes-by-cells expression matrix}

\item{dynRes}{result of running findDynGenes or define_epochs}

\item{num_epochs}{the number of epochs}

\item{limit_to}{vector of genes on which to base epoch cuts, for example, limiting to TFs}

\item{method}{what clustering method to use, either 'kmeans' or 'hierarchical'}

\item{pThresh_dyn}{pval threshold if gene is dynamically expressed}
}
\value{
vector of  pseudotimes at which to cut data into epochs
}
\description{
Returns cuts to define epochs via clustering
}
