% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{find_cuts_by_similarity}
\alias{find_cuts_by_similarity}
\title{Returns cuts to define epochs}
\usage{
find_cuts_by_similarity(expDat, dynRes, winSize = 5, pThresh_dyn = 0.05)
}
\arguments{
\item{expDat}{genes-by-cells expression matrix}

\item{dynRes}{individual path result of running define_epochs}

\item{winSize}{number of cells to each side to compare each cell to}

\item{pThresh_dyn}{pval threshold if gene is dynamically expressed}
}
\value{
vector of  pseudotimes at which to cut data into epochs
}
\description{
Returns cuts to define epochs
}
