% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_networkAnalysis.R
\name{find_targets}
\alias{find_targets}
\title{Finds binding targets given list of dataframes containing binding info for effectors}
\usage{
find_targets(
  aList,
  column = "max_score",
  by_rank = FALSE,
  n_targets = 2000,
  threshold = NULL
)
}
\arguments{
\item{aList}{the result of running score_targets}

\item{column}{column name used in either ranking or thresholding the possible targets}

\item{by_rank}{if TRUE will return top n_targets; if FALSE will use thresholds}

\item{n_targets}{the number of targets to return for each effector if by_rank=TRUE}

\item{threshold}{threshold to use if by_rank=FALSE}
}
\value{
list of binding targets for list of effectors
}
\description{
Finds binding targets given list of dataframes containing binding info for effectors
}
