% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.r
\name{get_config_data}
\alias{get_config_data}
\title{Get configuration data from remote URL}
\usage{
get_config_data(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical indicating whether to show messages}
}
\value{
A list of project configurations
}
\description{
Get configuration data from remote URL
}
