% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{grnKsmooth}
\alias{grnKsmooth}
\title{smooths expression}
\usage{
grnKsmooth(expDat, cells, BW = 0.25)
}
\arguments{
\item{expDat}{expression matrix}

\item{cells}{sample table that includes "cell_name"  "pseudotime" "group"}

\item{BW}{bandwith for kernel smoother}

\item{npoints}{n.points parameter for kernel smoother}
}
\value{
smoothed matrix
}
\description{
smooths expression across cells in path
}
