% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{hm_dyn}
\alias{hm_dyn}
\title{plots results of findDynGenes}
\usage{
hm_dyn(
  expDat,
  dynRes,
  topX = 25,
  cRow = FALSE,
  cCol = FALSE,
  limits = c(0, 10),
  toScale = FALSE,
  fontsize_row = 4,
  geneAnn = FALSE,
  anno_colors = NULL,
  show_rownames = TRUE
)
}
\arguments{
\item{expDat}{expression matrix}

\item{dynRes}{result of running findDynGenes}

\item{topX}{top x genes to plot
'}
}
\value{
heatmaps
}
\description{
heatmap
}
