% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{hm_dyn_clust}
\alias{hm_dyn_clust}
\title{plots results of findDynGenes}
\usage{
hm_dyn_clust(
  expDat,
  dynRes,
  geneAnn,
  row_cols,
  limits = c(0, 10),
  toScale = FALSE,
  fontsize_row = 4
)
}
\arguments{
\item{expDat}{expression matrix}

\item{dynRes}{result of running findDynGenes}

\item{topX}{top x genes to plot
'}
}
\value{
heatmaps
}
\description{
heatmap
}
