% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_networkAnalysis.R
\name{load_SP_effectors}
\alias{load_SP_effectors}
\title{Function to load in tsv files from Chip-Atlas}
\usage{
load_SP_effectors(path)
}
\arguments{
\item{path}{the path to the tsv files containing binding data per effector}
}
\value{
list of data frames
}
\description{
Function to load in tsv files from Chip-Atlas
}
