% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{loadDataFromLoom}
\alias{loadDataFromLoom}
\title{Loads .obs and .var data from loom file}
\usage{
loadDataFromLoom(
  path,
  col_attrs = NULL,
  obs_names = "obs_names",
  var_names = "var_names"
)
}
\arguments{
\item{path}{path to loom file}

\item{col_attrs}{column attributes to include in sample table. By default will include all in loom file.}

\item{obs_names}{where in loom file observation names are stored}

\item{var_names}{where in loom file variable names are stored}
}
\value{
list includeing expression matrix and sample table
}
\description{
Loads .obs and .var data from loom file
}
