% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_networkAnalysis.R
\name{mean_module_expression}
\alias{mean_module_expression}
\title{Computes mean expression of groups of genes}
\usage{
mean_module_expression(expX, module_list)
}
\arguments{
\item{expX}{genes-by-cells expression matrix}

\item{module_list}{list containing grouped genes, each element is a module of genes}
}
\value{
module expression
}
\description{
Computes mean expression of groups of genes
}
