% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_networkAnalysis.R
\name{mean_subnetwork_expression}
\alias{mean_subnetwork_expression}
\title{Computes mean expression of groups of genes in a dynamic network}
\usage{
mean_subnetwork_expression(expX, community_list)
}
\arguments{
\item{expX}{genes-by-cells expression matrix}

\item{community_list}{list of dataframes with community assginemnts. Each dataframe is the result of running subnets.}
}
\value{
subnetwork expression
}
\description{
Computes mean expression of groups of genes in a dynamic network
}
