% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_networkAnalysis.R
\name{order_genes}
\alias{order_genes}
\title{Function that orders genes based on peak expression}
\usage{
order_genes(exp, sampTab, pseudotime_column, smooth = TRUE)
}
\arguments{
\item{exp}{expression matrix}

\item{sampTab}{sample table}

\item{pseudotime_column}{column in sample table containing pseudotime annotation}

\item{smooth}{whether or not to smooth expression across pseudotime}
}
\value{
pheatmap
}
\description{
Function that orders genes based on peak expression
}
