% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{plot_dynamic_network}
\alias{plot_dynamic_network}
\title{quick plot of dynamic networks}
\usage{
plot_dynamic_network(grn, tfs, only_TFs = TRUE, order = NULL, thresh = NULL)
}
\arguments{
\item{grn}{the result of running epochGRN}

\item{tfs}{tfs}

\item{only_TFs}{plot only TF network}
}
\value{

}
\description{
quick plot of dynamic networks
}
