% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{plot_targets}
\alias{plot_targets}
\title{quick plot of top targets of specified regulators}
\usage{
plot_targets(
  grn,
  gene_ranks,
  tfs,
  tfstoplot = NULL,
  numTargets = 5,
  only_TFs = TRUE,
  order = NULL
)
}
\arguments{
\item{grn}{the result of running epochGRN}

\item{gene_ranks}{the result of running compute_pagerank}

\item{tfs}{tfs}

\item{tfstoplot}{named list of tfs to plot, matching order.}

\item{numTargets}{number of top targets to plot for each regulator}

\item{only_TFs}{plot only TF network}

\item{order}{which epochs or transitions to plot}
}
\value{

}
\description{
quick plot of top targets of specified regulators
}
