% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{plot_targets_and_regulators}
\alias{plot_targets_and_regulators}
\title{Updated plot of top regulators given targets in dynamic networks based on a weight column.
Top regulators computed for each epoch, but maintained in plot across epochs if present in epoch subnetwork.}
\usage{
plot_targets_and_regulators(
  grn,
  targets,
  epochs = NULL,
  weight_column = "zscore",
  gene_ranks = NULL,
  numTopRegulators = 5,
  order = NULL,
  fixed_layout = TRUE,
  declutter = TRUE,
  show_expression = TRUE,
  node_size = 5,
  label_size = 2,
  title_size = 10,
  legend_size = 8
)
}
\arguments{
\item{grn}{the result of running epochGRN}

\item{targets}{targets}

\item{epochs}{result of running assign_epochs}

\item{weight_column}{column name containing reconstruction weights to use}

\item{numTopRegulators}{number of top regulators to plot}

\item{order}{which epochs or transitions to plot}

\item{fixed_layout}{whether or not to fix node positions across epoch networks}

\item{declutter}{if TRUE, will only label nodes with active interactions in given network}

\item{show_expression}{if TRUE, size and shade of node indicates mean expression in a given epoch.}
}
\value{

}
\description{
Updated plot of top regulators given targets in dynamic networks based on a weight column.
Top regulators computed for each epoch, but maintained in plot across epochs if present in epoch subnetwork.
}
