% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{plot_targets_with_top_regulators}
\alias{plot_targets_with_top_regulators}
\title{quick plot of top regulators given targets in dynamic networks based on reconstruction weight}
\usage{
plot_targets_with_top_regulators(
  grn,
  targets,
  weight_column = "zscore",
  gene_ranks = NULL,
  numTopRegulators = 5,
  order = NULL
)
}
\arguments{
\item{grn}{the result of running epochGRN}

\item{targets}{targets}

\item{weight_column}{column name containing reconstruction weights to use}

\item{numTopRegulators}{number of top regulators to plot}

\item{order}{which epochs or transitions to plot}
}
\value{

}
\description{
quick plot of top regulators given targets in dynamic networks based on reconstruction weight
}
