% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{plot_targets_with_top_regulators_detail}
\alias{plot_targets_with_top_regulators_detail}
\title{quick plot of top regulators given targets in dynamic networks based on reconstruction weight, colored by expression and interaction type}
\usage{
plot_targets_with_top_regulators_detail(
  grn,
  targets,
  epochs,
  weight_column = "zscore",
  gene_ranks = NULL,
  numTopRegulators = 5,
  order = NULL,
  fixed_layout = TRUE,
  layout_alg = "fr",
  declutter = TRUE
)
}
\arguments{
\item{grn}{the result of running epochGRN}

\item{targets}{targets}

\item{epochs}{result of running assign_epochs}

\item{weight_column}{column name containing reconstruction weights to use}

\item{numTopRegulators}{number of top regulators to plot}

\item{order}{which epochs or transitions to plot}

\item{fixed_layout}{whether or not to fix node positions across epoch networks}

\item{layout_alg}{layout algorithm if fixed_layout. Defaults to FR. Ignored if fixed_layout==FALSE.}

\item{declutter}{if TRUE, will only label nodes with active interactions in given network}
}
\value{

}
\description{
quick plot of top regulators given targets in dynamic networks based on reconstruction weight, colored by expression and interaction type
}
