% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{plot_top_regulators}
\alias{plot_top_regulators}
\title{quick plot of top regulators in dynamic networks}
\usage{
plot_top_regulators(
  grn,
  gene_ranks,
  tfs,
  numTopTFs = 5,
  numTargets = 5,
  only_TFs = TRUE,
  order = NULL
)
}
\arguments{
\item{grn}{the result of running epochGRN}

\item{gene_ranks}{the result of running compute_pagerank}

\item{tfs}{tfs}

\item{numTopTFs}{number of top regulators to plot}

\item{numTargets}{number of top targets to plot for each regulator}

\item{only_TFs}{plot only TF network}

\item{order}{which epochs or transitions to plot}
}
\value{

}
\description{
quick plot of top regulators in dynamic networks
}
