% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_GRN.R
\name{reconstructGRN}
\alias{reconstructGRN}
\title{reconstruct GRN ala CLR}
\usage{
reconstructGRN(expDat, tfs, dgenes = NULL, method = "pearson", zThresh = 2)
}
\arguments{
\item{expDat}{unsmoothed expression matrix}

\item{tfs}{vector of transcription factor names}

\item{dgenes}{dynamic genes, found using findDynGenes}

\item{method}{either 'pearson' or 'MI' to be used in CLR}

\item{zThresh}{zscore threshold default 2}
}
\value{
data frame of TF TG zscore corr
}
\description{
reconstruct GRN ala CLR uses minet
}
