% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_GRN.R
\name{reconstructGRN_GENIE3}
\alias{reconstructGRN_GENIE3}
\title{reconstruct GRN ala GENIE3}
\usage{
reconstructGRN_GENIE3(expDat, tfs, dgenes = NULL, weightThresh = 0)
}
\arguments{
\item{expDat}{unsmoothed expression matrix}

\item{tfs}{vector of transcription factor names}

\item{dgenes}{dynamic genes, found using findDynGenes}

\item{weightThresh}{threshold to consider call positive default 0}
}
\value{
data frame of TF TG zscore corr
}
\description{
reconstruct GRN ala GENIE3 using GENIE3
}
