% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.r
\name{resample}
\alias{resample}
\alias{resample,Epoch-method}
\title{Generic function for resampling objects}
\usage{
resample(x, ...)

\S4method{resample}{Epoch}(x, freq, ...)
}
\arguments{
\item{x}{An \code{Epoch} object to be resampled.}

\item{...}{Additional arguments passed to \code{gsignal::resample}}

\item{freq}{The new sampling frequency.}
}
\value{
An \code{Epoch} object with the resampled data.
}
\description{
This function allows you to resample an object to a different sampling frequency.

This function allows you to resample an Epoch object to a different sampling frequency.
}
