% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkComparison.R
\name{sample_and_epoch_reconstruct}
\alias{sample_and_epoch_reconstruct}
\title{Function to bootstrap Epoch reconstruction}
\usage{
sample_and_epoch_reconstruct(
  expX,
  sampTab,
  tfs,
  n_reconstructions = 10,
  ncells_per_sample = 400,
  pThresh = 0.05,
  zThresh = 5,
  pseudotime_column = "dpt_pseudotime",
  group_column = "leiden",
  method = "pearson",
  crossweight = FALSE,
  limit_to = NULL
)
}
\arguments{
\item{expX}{genes-by-cells expression matrix}

\item{sampTab}{sample table}

\item{tfs}{TFs}

\item{n_reconstructions}{the number of networks to reconstruct}

\item{ncells_per_sample}{the number of cells to sample for each reconstruction}

\item{pThresh}{p-value threshold for selecting dynamic genes}

\item{zThresh}{z-score threshold in static reconstruction}

\item{pseudotime_column}{column name in sampTab with pseudotime annotation}

\item{group_column}{column name in sampTab for restricting reconstruction}

\item{method}{CLR method, either "pearson" or "MI"}

\item{crossweight}{whether or not to perform crossweighting}

\item{limit_to}{vector of genes. If not NULL, will skip finding dynamic genes and use genes in limit_to for reconstruction}
}
\value{
list of grnDFs
}
\description{
Function to bootstrap Epoch reconstruction
}
