% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_networkAnalysis.R
\name{score_targets}
\alias{score_targets}
\title{Function to score targets of effectors}
\usage{
score_targets(aList, threshold = 50)
}
\arguments{
\item{aList}{list of dataframes containing binding data for each effector}

\item{threshold}{binding score threshold to compute hit frequency}
}
\value{
updated list of dataframes
}
\description{
Adds columns for mean binding score, maximum binding score, and percent frequency target is a hit based on threshold
}
