% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-Epoch-print.r
\name{show,Epoch-method}
\alias{show,Epoch-method}
\title{Print the Epoch Object}
\usage{
\S4method{show}{Epoch}(object)
}
\arguments{
\item{object}{Epoch object}
}
\description{
Print the Epoch Object
}
\seealso{
Other Epoch methods: 
\code{\link{coltimes}()},
\code{\link{crop}()},
\code{\link{plot,Epoch,missing-method}},
\code{\link{resample}()}
}
\concept{Epoch methods}
