% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{split_epochs_by_group}
\alias{split_epochs_by_group}
\title{Splits data into epochs manually}
\usage{
split_epochs_by_group(dynRes, assignment)
}
\arguments{
\item{dynRes}{result of running findDynGenes or compileDynGenes}

\item{assignment}{a list of vectors where names(assignment) are epoch names, and vectors contain groups belonging to corresponding epoch}
}
\value{
updated dynRes with epoch column included in dynRes$cells
}
\description{
Splits data into epochs given group assignment
}
