% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{split_epochs_by_pseudotime}
\alias{split_epochs_by_pseudotime}
\title{Splits data into epochs}
\usage{
split_epochs_by_pseudotime(dynRes, cuts, epoch_names = NULL)
}
\arguments{
\item{dynRes}{result of running findDynGenes or compileDynGenes}

\item{cuts}{vector of pseudotime cutoffs}

\item{epoch_names}{names of resulting epochs, must have length of length(cuts)+1}
}
\value{
updated dynRes with epoch column included in dynRes$cells
}
\description{
Splits data into epochs by assigning cells to epochs
}
