% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_dynamicGRN.R
\name{tfnet_pagerank}
\alias{tfnet_pagerank}
\title{Function to compute page rank of TF-only networks}
\usage{
tfnet_pagerank(dynnet, tfs, weight_column = "zscore", directed_graph = FALSE)
}
\arguments{
\item{dynnet}{result of dynamic GRN reconstruction}

\item{tfs}{TFs}

\item{weight_column}{name of column in dynnet to weight edges}

\item{directed_graph}{if GRN is directed or not}
}
\value{
list of dataframes of active TFs in each epoch and transition, ranked by page rank
}
\description{
Function to compute page rank of TF-only networks
}
