% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting.R
\name{wn_ig_tabToIgraph}
\alias{wn_ig_tabToIgraph}
\title{convert a table to an igraph}
\usage{
wn_ig_tabToIgraph(
  grnTab,
  caoGenesRes,
  simplify = FALSE,
  directed = FALSE,
  weights = TRUE
)
}
\arguments{
\item{grnTab}{table of TF, TF, maybe zscores, maybe correlations}

\item{caoGenesRes}{result of running caoGenesRes data frame with gene peakTime, peakTimeRaw and epoch name}

\item{simplify}{false}

\item{directed}{FALSE,}

\item{weights}{TRUE}
}
\value{
iGraph object
}
\description{
convert a table to an igraph. This adds an nEnts vertex attribute to count the number of entities in the sub-net
}
