% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{.standardizeIEEG}
\alias{.standardizeIEEG}
\title{Standardize iEEG row data for plotting}
\usage{
.standardizeIEEG(data, standardize = TRUE, gap = 2)
}
\arguments{
\item{data}{Matrix or data frame of iEEG data}

\item{standardize}{Logical or numeric vector. If logical, indicates whether to standardize each row. If numeric, indicates the scaling factor for each row.}

\item{gap}{Numeric. The gap to separate different electrodes in the plot.}
}
\value{
Standardized data matrix
}
\description{
Standardize iEEG row data for plotting
}
