% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{.standardizeIEEG}
\alias{.standardizeIEEG}
\title{Standardize iEEG row data for plotting}
\usage{
.standardizeIEEG(data)
}
\arguments{
\item{data}{Matrix or data frame of iEEG data}
}
\value{
Standardized data matrix
}
\description{
Standardize iEEG row data for plotting
}
