% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Epoch.r, R/method-print.r
\name{crop}
\alias{crop}
\alias{crop,Epoch-method}
\alias{show,Epoch-method}
\title{Methods for Epoch class}
\usage{
crop(x, start, end)

\S4method{crop}{Epoch}(x, start, end)

\S4method{show}{Epoch}(object)
}
\arguments{
\item{start}{Numeric value specifying start of new time range}

\item{end}{Numeric value specifying end of new time range}

\item{object}{Epoch object}
}
\value{
clip: clip the time range of the Epoch object
}
\description{
\code{clip}: Truncating time range
}
