% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.r
\name{names,EpochDownloader-method}
\alias{names,EpochDownloader-method}
\alias{[,EpochDownloader-method}
\alias{$,EpochDownloader-method}
\alias{[[,EpochDownloader-method}
\alias{show,EpochDownloader-method}
\alias{length,EpochDownloader-method}
\title{EpochDownloader Methods}
\usage{
\S4method{names}{EpochDownloader}(x)

\S4method{[}{EpochDownloader}(x, i)

\S4method{$}{EpochDownloader}(x, name)

\S4method{[[}{EpochDownloader}(x, i)

\S4method{show}{EpochDownloader}(object)

\S4method{length}{EpochDownloader}(x)
}
\arguments{
\item{x}{An \code{EpochDownloader} object.}

\item{i}{Index or name of the files to be accessed.}

\item{name}{The name of the file to be accessed.}

\item{object}{An \code{EpochDownloader} object.}
}
\value{
\code{names}: A character vector of file names.

\code{[}: A list of \code{Epoch} objects

\code{$}: A single \code{Epoch} object.

\code{[[}: A single \code{Epoch} object.

\code{show}: Prints a summary of the \code{EpochDownloader} object.

\code{length}: Returns the number of files in the \code{EpochDownloader} object.
}
\description{
EpochDownloader Methods
}
