% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpochDownloader.r
\name{EpochDownloader}
\alias{EpochDownloader}
\title{EpochDownloader constructor}
\usage{
EpochDownloader(
  id = EpochRepos()[[1]],
  progress = TRUE,
  path = file.path(tempdir(), id)
)
}
\arguments{
\item{id}{Either the ID of the OSF project or the name of the iEEG dataset. Check a list of available projects using \code{EpochRepos()}. The default points to the fragility data from the Fragility multi-center retrospective study.}

\item{progress}{Logical indicating whether to show progress during download.}

\item{path}{The path to the temporary folder where the files will be downloaded.}
}
\value{
An \code{EpochDownloader} object.
}
\description{
EpochDownloader constructor
}
