% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.r
\name{EpochRepos}
\alias{EpochRepos}
\title{Get the list of available projects}
\usage{
EpochRepos()
}
\value{
A list of project names and their corresponding OSF project IDs.
}
\description{
This function returns the list of available projects.
If the configuration has not been fetched yet, it will
automatically update the project list from the remote repository.
}
