% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_GRN.R
\name{findDynGenes}
\alias{findDynGenes}
\title{find genes expressed dynamically}
\usage{
findDynGenes(
  expDat,
  sampTab,
  path = NULL,
  group_column = "dpt_groups",
  pseudotime_column = "pseudotime",
  method = "gam",
  expRaw = NULL
)
}
\arguments{
\item{expDat}{properly normalized expression matrix}

\item{sampTab}{sample table that includes pseudotime, rownames = cell_name, and a group column}

\item{path}{vector of group names to include}

\item{group_column}{column name in sampTab annotating groups in the path}

\item{pseudotime_column}{column name in sampTab annotating pseudotime or latent time}

\item{method}{method to find dynamic genes. Defaults to "gam", if "tradeseq", must provide expRaw}

\item{expRaw}{raw expression matrix, required if method="tradeseq"}
}
\value{
pvals and cell info
}
\description{
uses slingshot approach
}
