% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpochRepos.r
\name{get_config_data}
\alias{get_config_data}
\title{Get configuration data from remote URL}
\usage{
get_config_data()
}
\value{
A list of project configurations
}
\description{
Get configuration data from remote URL
}
